+++ esp-hal/src/parl_io.rs
@@ -586,12 +586,6 @@ tx_pins!(
     P15 = PARL_TX_DATA15
 );
 
-impl FullDuplex for TxOneBit<'_> {}
-impl FullDuplex for TxTwoBits<'_> {}
-impl FullDuplex for TxFourBits<'_> {}
-impl FullDuplex for TxEightBits<'_> {}
-impl FullDuplex for TxPinConfigWithValidPin<'_, TxFourBits<'_>> {}
-
 impl NotContainsValidSignalPin for TxOneBit<'_> {}
 impl NotContainsValidSignalPin for TxTwoBits<'_> {}
 impl NotContainsValidSignalPin for TxFourBits<'_> {}
@@ -792,12 +786,6 @@ rx_pins!(
     P15 = PARL_RX_DATA15
 );
 
-impl FullDuplex for RxOneBit<'_> {}
-impl FullDuplex for RxTwoBits<'_> {}
-impl FullDuplex for RxFourBits<'_> {}
-impl FullDuplex for RxEightBits<'_> {}
-impl FullDuplex for RxPinConfigWithValidPin<'_, RxFourBits<'_>> {}
-
 impl NotContainsValidSignalPin for RxOneBit<'_> {}
 impl NotContainsValidSignalPin for RxTwoBits<'_> {}
 impl NotContainsValidSignalPin for RxFourBits<'_> {}
@@ -825,7 +813,7 @@ where
         bit_order: BitPackOrder,
     ) -> Result<ParlIoTx<'d, Dm>, Error>
     where
-        P: FullDuplex + TxPins + ConfigurePins + 'd,
+        P: TxPins + ConfigurePins + 'd,
         CP: TxClkPin + 'd,
     {
         tx_pins.configure()?;
@@ -905,7 +893,7 @@ where
         timeout_ticks: Option<u16>,
     ) -> Result<ParlIoRx<'d, Dm>, Error>
     where
-        P: FullDuplex + RxPins + ConfigurePins,
+        P: RxPins + ConfigurePins,
         CP: RxClkPin,
     {
         let guard = GenericPeripheralGuard::new();
@@ -1860,8 +1848,6 @@ pub mod asynch {
 mod private {
     use super::{BitPackOrder, Error, SampleEdge};
     use crate::peripherals::PARL_IO;
-
-    pub trait FullDuplex {}
 
     pub trait NotContainsValidSignalPin {}